Add auth to, and test, collection APIs

+++ odp/api2/routers/collection.py
@@ -2,10 +2,11 @@ from typing import List
 
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy import select, func
-from starlette.status import HTTP_404_NOT_FOUND, HTTP_409_CONFLICT
+from starlette.status import HTTP_404_NOT_FOUND, HTTP_409_CONFLICT, HTTP_403_FORBIDDEN
 
+from odp import ODPScope
 from odp.api2.models import CollectionModelIn, CollectionModel, CollectionSort
-from odp.api2.routers import Pager, Paging
+from odp.api2.routers import Pager, Paging, Authorize, Authorized
 from odp.db import Session
 from odp.db.models import Collection, Record
 
@@ -16,7 +17,10 @@ router = APIRouter()
     '/',
     response_model=List[CollectionModel],
 )
-async def list_collections(pager: Pager = Depends(Paging(CollectionSort))):
+async def list_collections(
+        pager: Pager = Depends(Paging(CollectionSort)),
+        auth: Authorized = Depends(Authorize(ODPScope.COLLECTION_READ)),
+):
     stmt = (
         select(Collection, func.count(Record.id)).
         outerjoin(Record).
@@ -25,6 +29,8 @@ async def list_collections(pager: Pager = Depends(Paging(CollectionSort))):
         offset(pager.skip).
         limit(pager.limit)
     )
+    if auth.provider_ids != '*':
+        stmt = stmt.where(Collection.provider_id.in_(auth.provider_ids))
 
     collections = [
         CollectionModel(
@@ -46,6 +52,7 @@ async def list_collections(pager: Pager = Depends(Paging(CollectionSort))):
 )
 async def get_collection(
         collection_id: str,
+        auth: Authorized = Depends(Authorize(ODPScope.COLLECTION_READ)),
 ):
     stmt = (
         select(Collection, func.count(Record.id)).
@@ -56,6 +63,9 @@ async def get_collection(
 
     if not (result := Session.execute(stmt).one_or_none()):
         raise HTTPException(HTTP_404_NOT_FOUND)
+
+    if auth.provider_ids != '*' and result.Collection.provider_id not in auth.provider_ids:
+        raise HTTPException(HTTP_403_FORBIDDEN)
 
     return CollectionModel(
         id=result.Collection.id,
@@ -71,7 +81,11 @@ async def get_collection(
 )
 async def create_collection(
         collection_in: CollectionModelIn,
+        auth: Authorized = Depends(Authorize(ODPScope.COLLECTION_ADMIN)),
 ):
+    if auth.provider_ids != '*' and collection_in.provider_id not in auth.provider_ids:
+        raise HTTPException(HTTP_403_FORBIDDEN)
+
     if Session.get(Collection, collection_in.id):
         raise HTTPException(HTTP_409_CONFLICT)
 
@@ -88,7 +102,11 @@ async def create_collection(
 )
 async def update_collection(
         collection_in: CollectionModelIn,
+        auth: Authorized = Depends(Authorize(ODPScope.COLLECTION_ADMIN)),
 ):
+    if auth.provider_ids != '*' and collection_in.provider_id not in auth.provider_ids:
+        raise HTTPException(HTTP_403_FORBIDDEN)
+
     if not (collection := Session.get(Collection, collection_in.id)):
         raise HTTPException(HTTP_404_NOT_FOUND)
 
@@ -102,8 +120,12 @@ async def update_collection(
 )
 async def delete_collection(
         collection_id: str,
+        auth: Authorized = Depends(Authorize(ODPScope.COLLECTION_ADMIN)),
 ):
     if not (collection := Session.get(Collection, collection_id)):
         raise HTTPException(HTTP_404_NOT_FOUND)
+
+    if auth.provider_ids != '*' and collection.provider_id not in auth.provider_ids:
+        raise HTTPException(HTTP_403_FORBIDDEN)
 
     collection.delete()
